requirements.txt
django
djangorestframework
markdown
django-filter
djangorestframework-jsonp
djangorestframework-xml
djangorestframework-csv
django-crispy-forms

